use redis to sync create-destroy folder, file

+++ app/controllers/up_files_controller.rb
@@ -2,7 +2,7 @@ class UpFilesController < ApplicationController
     include ApplicationHelper
     include UpFilesHelper
 
-    before_action :set_up_file, only: [:show, :continue_upload, :edit, :rename, :update, :destroy]
+    before_action :set_up_file, only: [:continue_upload, :edit, :rename, :update, :destroy]
     before_action :set_new_items, only: [:show]
 
     # GET /up_files
@@ -12,7 +12,13 @@ class UpFilesController < ApplicationController
 
     # GET /up_files/1
     def show
-        @chat_room = @up_file.chat_room
+        begin
+            @up_file = UpFile.find(params[:id])
+            @chat_room = @up_file.chat_room
+        rescue ActiveRecord::RecordNotFound
+            flash[:warning] = 'Không tìm thấy file. Quay về home.'
+            redirect_to folder_path
+        end
     end
 
     # GET /up_files/new
@@ -36,6 +42,7 @@ class UpFilesController < ApplicationController
             @temp_upfile.status = 'ready'
             @temp_upfile.save
             save_direct_shortcut_for_file @temp_upfile, params[:folder_id]
+            UpFilesCreateBroadcastJob.set(wait: 2.second).perform_later current_folder.id
             return
         end
 
@@ -53,25 +60,35 @@ class UpFilesController < ApplicationController
         else
             @up_file = UpFile.find_by(temp_up_id: @temp_upfile.temp_up_id)
             @up_file.file_size += content_length
-            get_ftp_connection do |ftp|
-                full_path = ::File.dirname "#{ENV['ftp_folder']}/#{@up_file.link.path}"
-                base_name = File.basename(@up_file.link.to_s)
-                ftp.chdir(full_path)
-                ftp.getbinaryfile(base_name, base_name, 1024)
-                File.open(base_name, "ab") do |f|
+
+            if (Rails.env.development?)
+                #code for FTP server
+                get_ftp_connection do |ftp|
+                    full_path = ::File.dirname "#{ENV['ftp_folder']}/#{@up_file.link.path}"
+                    base_name = File.basename(@up_file.link.to_s)
+                    ftp.chdir(full_path)
+                    ftp.getbinaryfile(base_name, base_name, 1024)
+                    File.open(base_name, "ab") do |f|
+                        f.write(up_file_params[:link].read)
+                    end
+                    ftp.putbinaryfile(base_name, base_name, 1024)
+                    File.delete(base_name)
+                end
+            else
+                # code for local test
+                File.open(@up_file.link.path, "ab") do |f|
                     f.write(up_file_params[:link].read)
                 end
-                ftp.putbinaryfile(base_name, base_name, 1024)
-                File.delete(base_name)
             end
-
-
 
             if (end_of_chunk == total_chunk - 1)  # kết thúc
                 @up_file.status = 'ready'
             end
 
             @up_file.save
+            if @up_file.status == 'ready'
+                UpFilesCreateBroadcastJob.perform_later current_folder.id
+            end
         end
     end
 
@@ -90,7 +107,6 @@ class UpFilesController < ApplicationController
         @up_file.destroy
         redirect_to current_folder, notice: 'UpFile was successfully destroyed.'
         #render :json => true
-
     end
 
     def rename
@@ -100,6 +116,20 @@ class UpFilesController < ApplicationController
     def resume_upload
         @up_file = UpFile.find_by(temp_up_id: params[:id])
         render json: { file: {size: @up_file.file_size} } and return
+    end
+
+    def check_exist
+        respond_to do |format|
+            format.json {
+                begin
+                    @up_file = UpFile.find(params[:id])
+                    render json: true
+                rescue
+                    flash[:warning] = 'File không còn tồn tại.'
+                    render json: false
+                end
+            }
+        end
     end
 
     private
